outfitwar-signups/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
auraxium==0.2.2
backoff==2.1.2
certifi==2022.6.15
charset-normalizer==2.1.0
docopt==0.6.2
frozenlist==1.3.1
idna==3.3
multidict==6.0.2
pipreqs==0.4.11
pydantic==1.9.2
python-dateutil==2.8.2
pymongo==4.2.0
requests==2.28.1
typing_extensions==4.3.0
urllib3==1.26.11
websockets==10.3
yarg==0.1.9
yarl==1.8.1
